orchestrator/scripts/pdf-bridge/requirements.txt
PyMuPDF>=1.23.0,<2.0.0
pdfminer.six>=20221105
python-dateutil>=2.8.0
